refactor: condense internal packages to one
Was consistently fight import loops because I didn't know what I was
doing when I laid this project out. This still doesn't feel very
golang-y, but it's a better start.
