--- internal/bootstrap/github_release_test.go
@@ -1,1 +1,0 @@
-package bootstrap
