--- internal/bootstrap/interface.go
@@ -1,11 +1,0 @@
-package bootstrap
-
-type Item interface {
-	Check() (bool, error)
-	Install() error
-}
-
-//type Runner interface {
-//    RunAll([]Item) error
-//    RunSingle(Item) error
-//}
