+++ internal/lib/builder.go
@@ -1,4 +1,4 @@
-package builder
+package lib
 
 import (
 	"errors"
@@ -8,14 +8,12 @@ import (
 	"path/filepath"
 
 	"github.com/hashicorp/go-multierror"
-	"github.com/nicjohnson145/godot/internal/config"
-	"github.com/nicjohnson145/godot/internal/file"
 	"github.com/nicjohnson145/godot/internal/util"
 )
 
 type Builder struct {
 	Getter util.HomeDirGetter
-	Config *config.Config
+	Config *Config
 }
 
 func (b *Builder) Build(force bool) error {
@@ -55,10 +53,10 @@ func (b *Builder) Build(force bool) error {
 	return errs.ErrorOrNil()
 }
 
-func (b *Builder) buildFileObjs(m config.StringMap) []file.File {
-	files := make([]file.File, 0, len(m))
+func (b *Builder) buildFileObjs(m StringMap) []File {
+	files := make([]File, 0, len(m))
 	for tmpl, dest := range m {
-		files = append(files, file.File{
+		files = append(files, File{
 			TemplatePath:    filepath.Join(b.Config.TemplateRoot, tmpl),
 			DestinationPath: util.ReplacePrefix(dest, "~/", b.Config.Home),
 		})
@@ -66,8 +64,8 @@ func (b *Builder) buildFileObjs(m config.StringMap) []file.File {
 	return files
 }
 
-func (b *Builder) makeTemplateVars() file.TemplateVars {
-	return file.TemplateVars{
+func (b *Builder) makeTemplateVars() TemplateVars {
+	return TemplateVars{
 		Target:     b.Config.Target,
 		Submodules: filepath.Join(b.Config.DotfilesRoot, "submodules"),
 		Home:       b.Config.Home,
@@ -76,7 +74,7 @@ func (b *Builder) makeTemplateVars() file.TemplateVars {
 
 func (b *Builder) ensureConfig() {
 	if b.Config == nil {
-		b.Config = config.NewConfig(b.Getter)
+		b.Config = NewConfig(b.Getter)
 	}
 }
