+++ internal/lib/repo.go
@@ -1,4 +1,4 @@
-package repo
+package lib
 
 import (
 	"bytes"
