+++ internal/lib/repo_item.go
@@ -1,4 +1,4 @@
-package bootstrap
+package lib
 
 import (
 	"fmt"
