+++ internal/lib/untar.go
@@ -1,4 +1,4 @@
-package bootstrap
+package lib
 
 import (
 	"archive/tar"
