+++ main.go
@@ -4,12 +4,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/nicjohnson145/godot/internal/bootstrap"
-	"github.com/nicjohnson145/godot/internal/builder"
-	"github.com/nicjohnson145/godot/internal/config"
-	"github.com/nicjohnson145/godot/internal/controller"
-	"github.com/nicjohnson145/godot/internal/repo"
 	"github.com/nicjohnson145/godot/internal/util"
+	"github.com/nicjohnson145/godot/internal/lib"
 	"github.com/spf13/cobra"
 )
 
@@ -22,10 +18,10 @@ func main() {
 
 type Components struct {
 	HomeDirGetter util.HomeDirGetter
-	Repo          repo.Repo
-	Config        *config.Config
-	Builder       *builder.Builder
-	Runner        controller.ItemRunner
+	Repo          lib.Repo
+	Config        *lib.Config
+	Builder       *lib.Builder
+	Runner        lib.ItemRunner
 }
 
 type Main struct {
@@ -85,19 +81,19 @@ func New(dependencies Components) Main {
 		m.dependencies.HomeDirGetter = &util.OSHomeDir{}
 	}
 	if m.dependencies.Config == nil {
-		m.dependencies.Config = config.NewConfig(m.dependencies.HomeDirGetter)
+		m.dependencies.Config = lib.NewConfig(m.dependencies.HomeDirGetter)
 	}
 	if m.dependencies.Builder == nil {
-		m.dependencies.Builder = &builder.Builder{
+		m.dependencies.Builder = &lib.Builder{
 			Getter: m.dependencies.HomeDirGetter,
 			Config: m.dependencies.Config,
 		}
 	}
 	if m.dependencies.Runner == nil {
-		m.dependencies.Runner = bootstrap.NewRunner()
+		m.dependencies.Runner = lib.NewRunner()
 	}
 	if m.dependencies.Repo == nil {
-		m.dependencies.Repo = repo.NewShellGitRepo(m.dependencies.Config.DotfilesRoot)
+		m.dependencies.Repo = lib.NewShellGitRepo(m.dependencies.Config.DotfilesRoot)
 	}
 
 	// Root cmd
@@ -125,7 +121,7 @@ func New(dependencies Components) Main {
 		"Apply the command to the given target, not supplying a value (i.e --target vs --target=a), will result in the current target being used. "+
 			"The special value of 'ALL' will apply the change to all available targets",
 	)
-	m.rootCmd.PersistentFlags().Lookup("target").NoOptDefVal = config.CURRENT
+	m.rootCmd.PersistentFlags().Lookup("target").NoOptDefVal = lib.CURRENT
 
 	// Cease Cmds
 	m.ceaseCmd = &cobra.Command{
@@ -139,7 +135,7 @@ func New(dependencies Components) Main {
 			"open. If no target is given, the current target will be used",
 		Args: cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.TargetRemoveFile(m.target, args)
 		},
 	}
@@ -150,7 +146,7 @@ func New(dependencies Components) Main {
 			"selection prompt wil open. If no target is given, the current target will be used",
 		Args: cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.RemoveTargetBootstrap(m.target, args)
 		},
 	}
@@ -162,7 +158,7 @@ func New(dependencies Components) Main {
 			"will open. If no target is given, the current target will be used.",
 		Args: cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.TargetCeaseGithubRelease(m.target, args)
 		},
 	}
@@ -178,8 +174,8 @@ func New(dependencies Components) Main {
 		Short: "Build and symlink dotfiles",
 		Long:  "Compile templates, symlink them to their final destinations, install configured bootstraps",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
-			return c.Sync(controller.SyncOpts{Force: m.force, NoBootstrap: m.noBootstrap})
+			c := lib.NewController(m.controllerOpts())
+			return c.Sync(lib.SyncOpts{Force: m.force, NoBootstrap: m.noBootstrap})
 		},
 	}
 	// Sync Flags
@@ -197,7 +193,7 @@ func New(dependencies Components) Main {
 		Short: "List files managed by godot",
 		Long:  "List files managed by godot, if -t/--target is given, only that targets files will be listed",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.ShowFilesEntry(m.target, cmd.OutOrStdout())
 		},
 	}
@@ -206,7 +202,7 @@ func New(dependencies Components) Main {
 		Short: "List bootstrap items managed by godot",
 		Long:  "List bootstrap items managed by godot, if -t/--target is given, only that targets bootstrap items will be listed",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.ShowBootstrapsEntry(m.target, cmd.OutOrStdout())
 		},
 	}
@@ -216,7 +212,7 @@ func New(dependencies Components) Main {
 		Short:   "List github releases managed by godot",
 		Long:    "List github releases managed by godot, if -t/--target is given, only that targets in-use releases items will be listed",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.ShowGithubRelease(m.target, cmd.OutOrStdout())
 		},
 	}
@@ -233,8 +229,8 @@ func New(dependencies Components) Main {
 		Long:  "Edit a file managed by godot, if no file is suppled a prompt will be displayed",
 		Args:  cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
-			return c.EditFile(args, controller.EditOpts{NoSync: m.noSync})
+			c := lib.NewController(m.controllerOpts())
+			return c.EditFile(args, lib.EditOpts{NoSync: m.noSync})
 		},
 	}
 	m.editCmd.Flags().BoolVarP(&m.noSync, "no-sync", "s", false, "Do not issue a sync after editing")
@@ -251,7 +247,7 @@ func New(dependencies Components) Main {
 			"open. If no target is given, the current target will be used",
 		Args: cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.TargetAddFile(m.target, args)
 		},
 	}
@@ -262,7 +258,7 @@ func New(dependencies Components) Main {
 			"selection prompt wil open. If no target is given, the current target will be used",
 		Args: cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.AddTargetBootstrap(m.target, args)
 		},
 	}
@@ -273,7 +269,7 @@ func New(dependencies Components) Main {
 		Long:    "Add a github release to a target. If no target is given, the current target will be used.",
 		Args:    cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.TargetUseGithubRelease(m.target, args[0], m.location, m.trackUpdates)
 		},
 	}
@@ -308,7 +304,7 @@ func New(dependencies Components) Main {
 		Short: "Add a file",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.Import(args[0], m.as)
 		},
 	}
@@ -317,7 +313,7 @@ func New(dependencies Components) Main {
 		Short: "Add a bootstrap item",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
-			c := controller.NewController(m.controllerOpts())
+			c := lib.NewController(m.controllerOpts())
 			return c.AddBootstrapItem(args[0], m.manager, m.name, m.location)
 		},
 	}
@@ -344,7 +340,7 @@ func New(dependencies Components) Main {
 		"manager",
 		"m",
 		"",
-		"The name of the package manager, one of "+strings.Join(config.ValidManagers, ", "),
+		"The name of the package manager, one of "+strings.Join(lib.ValidManagers, ", "),
 	)
 	m.addBootstrapCmd.MarkFlagRequired("manager")
 	m.addBootstrapCmd.Flags().StringVarP(
@@ -367,8 +363,8 @@ func New(dependencies Components) Main {
 	return m
 }
 
-func (m Main) controllerOpts() controller.ControllerOpts {
-	return controller.ControllerOpts{
+func (m Main) controllerOpts() lib.ControllerOpts {
+	return lib.ControllerOpts{
 		HomeDirGetter: m.dependencies.HomeDirGetter,
 		Repo:          m.dependencies.Repo,
 		Config:        m.dependencies.Config,
